[QuangVV - 2024-03-30] Format Code, Update VnPay, Customer Order List
