+++ BusinessObject/Model/VnPaymentResponseModel.cs
@@ -1,0 +1,23 @@
+﻿namespace BusinessObject.Model
+{
+    public class VnPaymentResponseModel
+    {
+        public bool Success { get; set; }
+        public string PaymentMethod { get; set; }
+        public string OrderDescription { get; set; }
+        public string OrderId { get; set; }
+        public string PaymentId { get; set; }
+        public string TransactionId { get; set; }
+        public string Token { get; set; }
+        public string VnPayResponseCode { get; set; }
+    }
+
+    public class VnPaymentRequestModel
+    {
+        public int OrderId { get; set; }
+        public string Description { get; set; }
+        public double Amount { get; set; }
+        public DateTime CreateDate { get; set; }
+        public Order? Order { get; set; }
+    }
+}
