+++ WebApp/Services/Interface/IVnPayService.cs
@@ -1,0 +1,10 @@
+﻿using BusinessObject.Model;
+
+namespace WebApp.Services.Interface
+{
+    public interface IVnPayService
+    {
+        string CreatePaymentUrl(HttpContext context, VnPaymentRequestModel requestModel);
+        VnPaymentResponseModel PaymentExecute(IQueryCollection query);
+    }
+}
