+++ WebApp/Services/VNPayService.cs
@@ -1,0 +1,81 @@
+﻿using BusinessObject.Model;
+using Ultility;
+using WebApp.Services.Interface;
+
+namespace WebApp.Services
+{
+    public class VNPayService : IVnPayService
+    {
+        private readonly IConfiguration _config;
+
+        public VNPayService(IConfiguration config)
+        {
+            _config = config;
+        }
+
+        public string CreatePaymentUrl(HttpContext context, VnPaymentRequestModel requestModel)
+        {
+            var tick = DateTime.Now.Ticks.ToString();
+
+            var vnpay = new VnPayLibrary();
+            vnpay.AddRequestData("vnp_Version", _config["VnPay:Version"]);
+            vnpay.AddRequestData("vnp_Command", _config["VnPay:Command"]);
+            vnpay.AddRequestData("vnp_TmnCode", _config["VnPay:TmnCode"]);
+            vnpay.AddRequestData("vnp_Amount", (requestModel.Amount * 100).ToString());
+            vnpay.AddRequestData(
+                "vnp_CreateDate",
+                requestModel.CreateDate.ToString("yyyyMMddHHmmss")
+            );
+            vnpay.AddRequestData("vnp_CurrCode", _config["VnPay:CurrCode"]);
+            vnpay.AddRequestData("vnp_IpAddr", Utils.GetIpAddress(context));
+            vnpay.AddRequestData("vnp_Locale", _config["VnPay:Locale"]);
+            vnpay.AddRequestData("vnp_OrderInfo", requestModel.Order?.Id.ToString() ?? "0");
+            vnpay.AddRequestData("vnp_OrderType", "other");
+            vnpay.AddRequestData("vnp_ReturnUrl", _config["VnPay:PaymentBackReturnUrl"]);
+            vnpay.AddRequestData("vnp_TxnRef", tick);
+            var paymentUrl = vnpay.CreateRequestUrl(
+                _config["VnPay:BaseUrl"],
+                _config["VnPay:HashSecret"]
+            );
+
+            return paymentUrl;
+        }
+
+        public VnPaymentResponseModel PaymentExecute(IQueryCollection query)
+        {
+            var vnpay = new VnPayLibrary();
+            foreach (var (key, value) in query)
+            {
+                if (!string.IsNullOrEmpty(key) && key.StartsWith("vnp_"))
+                {
+                    vnpay.AddResponseData(key, value.ToString());
+                }
+            }
+            var vnp_OrderId = Convert.ToInt64(vnpay.GetResponseData("vnp_TxnRef"));
+            var vnp_TransactionId = Convert.ToInt64(vnpay.GetResponseData("vnp_TransactionNo"));
+            var vnp_SecureHash = query.FirstOrDefault(p => p.Key == "vnp_SecureHash").Value;
+            var vnp_ResponseCode = vnpay.GetResponseData("vnp_ResponseCode");
+            var vnp_OrderInfo = vnpay.GetResponseData("vnp_OrderInfo");
+
+            bool checkSignature = vnpay.ValidateSignature(
+                vnp_SecureHash,
+                _config["VnPay:HashSecret"]
+            );
+            if (!checkSignature)
+            {
+                return new VnPaymentResponseModel { Success = false };
+            }
+
+            return new VnPaymentResponseModel
+            {
+                Success = true,
+                PaymentMethod = "VnPay",
+                OrderDescription = vnp_OrderInfo,
+                OrderId = vnp_OrderId.ToString(),
+                TransactionId = vnp_TransactionId.ToString(),
+                Token = vnp_SecureHash,
+                VnPayResponseCode = vnp_ResponseCode,
+            };
+        }
+    }
+}
