+++ WebApp/Views/Customer/Order.cshtml
@@ -1,73 +1,106 @@
-﻿@model BusinessObject.Model.Order
-<div class="container mt-5">
-    <h2 class="mb-4">Order Details</h2>
-    <div class="row">
-        <div class="col-md-6">
-            <h4>Order Information</h4>
-            <p><strong>Order ID:</strong> @Model.Id</p>
-            <p><strong>Order Date:</strong> @Model.OrderDate.ToString("dd/MM/yyyy HH:mm")</p>
-            <p><strong>Status:</strong> @Model.Status</p>
-            <p><strong>Payment Method:</strong> @Model.PaymentMethod</p>
-            <p><strong>Shipping Address:</strong> @Model.ShippingAddress</p>
-            <p><strong>Phone Number:</strong> @Model.PhoneNumber</p>
-
-            @foreach (var detail in Model.OrderDetails)
-            {
-                <div class="card mb-3">
-                    <div class="card-body">
-                        <h5>@detail.Product.Name</h5>
-                        <p>Quantity: @detail.Quantity</p>
-                        <p>Price: $@(detail.UnitPrice * detail.Quantity)</p>
-                    </div>
-                </div>
-            }
-        </div>
-        <div class="col-md-6">
-            <h4>Summary</h4>
+﻿@model Order
+<div class="container mt-4 mb-4">
+    <div class="row d-flex align-items-center justify-content-center">
+        <div class="col-md-10">
             <div class="card">
-                <div class="card-body">
-                    <div class="d-flex justify-content-between">
-                        <strong>Total Price:</strong>
-                        <strong>$@Model.TotalPrice</strong>
+                <div class="row">
+                    <!-- Left Column - Order Information -->
+                    <div class="col-md-6 p-4">
+                        <h4>Order Information</h4>
+                        <div class="order-info-section">
+                            <p><strong>Order ID:</strong> <span>@Model.Id</span></p>
+                            <p><strong>Order Date:</strong> <span>@Model.OrderDate.ToString("yyyy/MM/dd")</span></p>
+                            <p><strong>Status:</strong> <span class="badge bg-danger">@Model.Status</span></p>
+                            <p><strong>Payment Method:</strong> <span>@Model.PaymentMethod</span></p>
+                            <p><strong>Shipping Address:</strong> <span>@Model.ShippingAddress</span></p>
+                            <p><strong>Phone Number:</strong> <span>@Model.PhoneNumber</span></p>
+                        </div>
+
+
+                        <h4>Order Items</h4>
+                        <div id="orderItems">
+                            <!-- This section will be populated with order items from the model -->
+                            <!-- Example item structure -->
+                            @if(Model.OrderDetails != null)
+                            {
+                                @foreach (var item in Model.OrderDetails)
+                                {
+                                    <div class="product-card p-3">
+                                        <div class="d-flex justify-content-between">
+                                            <h5>@item.Product?.Name</h5>
+                                            <span>
+                                                @item.Product?.Price.ToString("C", new CultureInfo("vi-VN"))
+                                            </span>
+                                        </div>
+                                        <p class="mb-0">Quantity: @item.Quantity</p>
+                                    </div>
+                                }
+                            }                            
+                        </div>
+                    </div>
+
+                    <!-- Right Column - Order Summary -->
+                    <div class="col-md-6 p-4 bg-light">
+                        <h4>Order Summary</h4>
+                        <div class="summary-card">
+                            <div class="summary-item">
+                                <strong>Subtotal</strong>
+                                <span>@Model.TotalPrice.ToString("C", new CultureInfo("vi-VN"))</span>
+                            </div>
+                            <div class="summary-item summary-total">
+                                <strong>Total</strong>
+                                <span>@Model.TotalPrice.ToString("C", new CultureInfo("vi-VN"))</span>
+                            </div>
+                        </div>
+
+                        <div class="mt-4">
+                            <div class="d-flex justify-content-between align-items-center mb-3">
+                                <span><i class="far fa-clock text-muted"></i> 1 hour delivery</span>
+                            </div>
+
+                            <a asp-action="Index" asp-controller="Home" class="btn btn-primary mt-2 w-100">Back to Home</a>
+                            <a asp-action="OrderHistory" asp-controller="Customer" class="btn btn-outline-secondary mt-2 w-100">View Order History</a>
+                        </div>
                     </div>
                 </div>
             </div>
-            @if (Model.Status == OrderStatus.Pending)
-            {
-                <button id="cancelOrder" class="btn btn-danger mt-3" data-order-id="@Model.Id">Cancel Order</button>
-            }
         </div>
     </div>
 </div>
 
-@section Scripts {
-    <script>
-        document.getElementById('cancelOrder')?.addEventListener('click', async () => {
-            const orderId = document.getElementById('cancelOrder').dataset.orderId;
-            const confirmCancel = confirm("Bạn có chắc chắn muốn hủy đơn hàng này không?");
+<style >
+    .summary-item {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+    }
 
-            if (confirmCancel) {
-                const response = await fetch(`/Customer/CancelOrder?orderId=${orderId}`, {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' }
-                });
-                const result = await response.json();
-                if (result.success) {
-                    alert(result.message);
-                    location.reload();
-                } else {
-                    alert(result.message);
-                }
-            } 
-        });
+        .summary-item span {
+            margin-left: auto;
+        }
 
-        @if (TempData["success"] != null)
-        {
-            @:alert('@TempData["success"]');
-        }
-        else if (TempData["error"] != null)
-        {
-            @:alert('@TempData["error"]');
-        }
-    </script>
-}
+    .badge {
+        max-width: 70px;
+        display: inline-block;
+        padding: 5px 10px;
+        border-radius: 5px;
+        text-align: center;
+        white-space: nowrap;
+    }
+    .order-info-section p {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        max-width: 500px; 
+    }
+
+    .order-info-section strong {
+        min-width: 150px; 
+    }
+
+    .order-info-section span {
+        flex: 1;
+        text-align: right; 
+    }
+</style>
+
