--- WebApp/Views/Customer/ProductDetail.cshtml
@@ -1,46 +1,0 @@
-﻿@model BusinessObject.Model.Product
-
-<div class="container mt-5">
-    <div class="row">
-        <div class="col-md-6">
-            @if (Model.ProductAvatar != null)
-            {
-                <img src="@Url.Content(Model.ProductAvatar.ImagePath)" alt="@Model.Name" class="img-fluid" style="max-height: 400px; object-fit: contain;" />
-            }
-            <div class="mt-3">
-                <h5>Ảnh khác</h5>
-                <div class="d-flex flex-wrap">
-                    @foreach (var image in Model.ProductImages)
-                    {
-                        <img src="@Url.Content(image.ImagePath)" alt="Product Image" class="img-thumbnail me-2 mb-2" style="max-width: 100px;" />
-                    }
-                </div>
-            </div>
-        </div>
-        <div class="col-md-6">
-            <h2>@Model.Name</h2>
-            <p class="text-muted">Danh mục: @Model.Category?.Name</p>
-            <h4 class="text-primary">$@Model.Price</h4>
-            <p>@Model.Description</p>
-            <a asp-action="AddToCart" asp-controller="Cart" asp-route-productId="@Model.Id" class="btn btn-primary mt-3">Thêm vào giỏ hàng</a>
-        </div>
-    </div>
-    <div class="mt-5">
-        <h3>Feedback</h3>
-        @if (Model.Feedbacks != null && Model.Feedbacks.Any())
-        {
-            <ul class="list-group">
-                @foreach (var feedback in Model.Feedbacks)
-                {
-                    <li class="list-group-item">
-                        <strong>@feedback.User?.UserName</strong>: @feedback.FeedbackStars sao - @feedback.FeedbackContent
-                    </li>
-                }
-            </ul>
-        }
-        else
-        {
-            <p>Chưa có feedback nào cho sản phẩm này.</p>
-        }
-    </div>
-</div>
