+++ WebApp/Views/Home/Menu.cshtml
@@ -77,7 +77,9 @@
                                                 <div class="w-100 d-flex flex-column text-start ps-4">
                                                     <h5 class="d-flex justify-content-between border-bottom pb-2">
                                                         <span>@food.Name</span>
-                                                        <span class="text-primary">@food.Price.ToString("C", new CultureInfo("en-US"))</span>
+                                                <span class="text-primary">
+                                                    @food.Price.ToString("C", new CultureInfo("vi-VN"))
+                                                </span>
                                                     </h5>
                                                     @if (!string.IsNullOrEmpty(food.Description))
                                                     {
@@ -110,7 +112,7 @@
                                                 <div class="w-100 d-flex flex-column text-start ps-4">
                                                     <h5 class="d-flex justify-content-between border-bottom pb-2">
                                                         <span>@drink.Name</span>
-                                                        <span class="text-primary">@drink.Price.ToString("C", new CultureInfo("en-US"))</span>
+                                                        <span class="text-primary">@drink.Price.ToString("C", new CultureInfo("vi-VN"))</span>
                                                     </h5>
                                                     @if (!string.IsNullOrEmpty(drink.Description))
                                                     {
@@ -143,7 +145,7 @@
                                                 <div class="w-100 d-flex flex-column text-start ps-4">
                                                     <h5 class="d-flex justify-content-between border-bottom pb-2">
                                                         <span>@combo.Name</span>
-                                                        <span class="text-primary">@combo.Price.ToString("C", new CultureInfo("en-US"))</span>
+                                                        <span class="text-primary">@combo.Price.ToString("C", new CultureInfo("vi-VN"))</span>
                                                     </h5>
                                                     @if (!string.IsNullOrEmpty(combo.Description))
                                                     {
