+++ WebApp/Views/Home/ProductDetail.cshtml
@@ -100,7 +100,9 @@
                     </div>
 
                     <div class="mb-3">
-                        <span class="h5">@Model.Price.ToString("C", new CultureInfo("en-US"))</span>
+                        <span class="h5">
+                            @Model.Price.ToString("C", new CultureInfo("vi-VN"))
+                        </span>
                     </div>
 
                     <dl class="row">
@@ -109,12 +111,7 @@
 
                     </dl>
                     <hr />
-                    <a asp-action="" asp-controller="" class="btn btn-success shadow-0"> Buy now </a>
-                    <a asp-action="" asp-controller="" class="btn btn-primary shadow-0"> <i class="me-1 fa fa-shopping-basket"></i> Add to cart </a>
-                    @if (Model != null)
-                    {
-                        <partial name="_Feedback" model="Model" />
-                    }
+                    <a asp-action="AddToCart" asp-controller="Customer" asp-route-productId="@Model.Id" class="btn btn-primary shadow-0"> <i class="me-1 fa fa-shopping-basket"></i> Add to cart </a>
                 </div>
             </main> 
         </div>
