+++ WebApp/Views/Shared/_Feedback.cshtml
@@ -1,6 +1,6 @@
 ﻿
 <link rel="stylesheet" href="~/css/feedback.css">
-<button id="btnFeedback" class="btn btn-primary shadow-0" data-bs-toggle="modal" data-bs-target="#feedbackModal">Feedback</button>
+<button id="btnFeedback" class="btn btn-outline-primary shadow-0" data-bs-toggle="modal" data-bs-target="#feedbackModal">Feedback</button>
 <!-- Modal Feedback  -->
 <div class="modal fade" id="feedbackModal" tabindex="-1" aria-labelledby="feedbackModalLabel" aria-hidden="true">
     <div class="modal-dialog modal-lg w-100">
