+++ WebApp/Views/Shared/_NormalNavbar.cshtml
@@ -21,7 +21,7 @@
                 @if (User.Identity.IsAuthenticated)
                 {
                     <li class="nav-item me-3">
-                        <a class="d-block" asp-area="" asp-controller="Customer" asp-action="" title="Shopping Cart">
+                        <a class="d-block" asp-area="" asp-controller="Customer" asp-action="CartIndex" title="Shopping Cart">
                             <div class="rounded-circle bg-light d-flex align-items-center justify-content-center" style="width: 38px; height: 38px;">
                                 <i class="fa-solid fa-cart-shopping"></i>
                             </div>
