+++ WebApp/appsettings.json
@@ -13,7 +13,17 @@
     "SmtpUser": "quangvvde170166@fpt.edu.vn",
     "SmtpPass": "deeqgtkupmrkvcvv"
   },
+  "VnPay": {
+    "TmnCode": "JF3IY54S",
+    "HashSecret": "QZOTBZCKSA6E0UYRUUBTER6XHCIN15QX",
+    "BaseUrl": "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html",
+    "Version": "2.1.0",
+    "Command": "pay",
+    "CurrCode": "VND",
+    "Locale": "vn",
+    "PaymentBackReturnUrl": "https://localhost:7129/Customer/VNPayReturn"
 
+  },
   "GoogleKeys": {
     "ClientID": "784925800619-1g3lgienm08f1mlk47pc81dpclpa7fd2.apps.googleusercontent.com",
     "ClientSecret": "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"
